Code cleanup and code restructure for disk stats collection and submition
