+++ internal/lotus/services/lotus_service.go
@@ -45,6 +45,7 @@ func (ls *lotusService) GetMinerAddress() (string, error) {
 		return "", err
 	}
 	return processResult(response)
+	// return "miner-address", nil
 }
 
 func processResult(response *jsonrpc.RPCResponse) (string, error) {
