+++ internal/stats/stats.go
@@ -1,12 +1,11 @@
 package stats
 
 import (
-	"fmt"
 	"github.com/NodeFactoryIo/hactar-daemon/internal/hactar"
 	"github.com/NodeFactoryIo/hactar-daemon/internal/lotus/services"
+	"github.com/NodeFactoryIo/hactar-daemon/internal/url"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
-	"net/http"
 	"strconv"
 	"time"
 )
@@ -15,36 +14,18 @@ func SubmitNewStatsReport() bool  {
 	client := hactar.NewClient(nil)
 	lotusService := services.NewLotusService(nil, nil)
 
+	nodeUrl := url.GetUrl()
 	actorAddress, err := lotusService.GetMinerAddress()
 	if err != nil {
-		fmt.Print("Worker down!")
-		// return nil
+		log.Error("Unable to send stats report because worker is down.")
+		return false
 	}
 
+	// send all statistics
+	log.Info("Collecting stats and sending report.")
+	SendDiskInfoStats(client, actorAddress, nodeUrl)
 
-	sendDiskInfoStats(client, actorAddress, "")
-	// TODO collect all statistic and send it to backend
-	fmt.Println("Collecting stats and sending report....")
 	return true
-}
-
-func sendDiskInfoStats(client *hactar.Client, actorAddress string, nodeUrl string) {
-	usage := DiskUsage("/")
-
-	response, err := client.DiskInfo.SendDiskInfo(hactar.DiskInfo{
-		FreeSpace:    string(usage.Free),
-		TakenSpace:   string(usage.Used),
-		NodeUrl:      nodeUrl,
-		ActorAddress: actorAddress,
-	})
-
-	if err != nil {
-		log.Error("Unable to send disk information statistics. ", err)
-	}
-
-	if response != nil && response.StatusCode == http.StatusOK {
-		log.Info(fmt.Sprintf("Disk: %d (free) %d (used)", usage.Free, usage.Used))
-	}
 }
 
 func StartMonitoringStats() {
@@ -58,7 +39,7 @@ func StartMonitoringStats() {
 			case <-done:
 				return
 			case <-ticker.C:
-				log.Info("Stats monitor ticked")
+				log.Info("Stats monitor ticked.")
 				SubmitNewStatsReport()
 			}
 		}
