+++ pkg/util/util.go
@@ -2,8 +2,6 @@ package util
 
 import (
 	"fmt"
-	log "github.com/sirupsen/logrus"
-	"net"
 )
 
 // Panic if error exists
@@ -16,19 +14,4 @@ func Must(err error, message string) {
 // Converts interface to string
 func String(i interface{}) string {
 	return fmt.Sprintf("%v", i)
-}
-
-// Get preferred outbound ip of this machine
-func GetOutboundIP() net.IP {
-	// get the local up address if it would connect to that target,
-	// can be changed to any other IP address.
-	conn, err := net.Dial("udp", "8.8.8.8:80") // this
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer conn.Close()
-
-	localAddr := conn.LocalAddr().(*net.UDPAddr)
-
-	return localAddr.IP
 }
